requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
python-dotenv==1.0.0
requests==2.31.0
Werkzeug==2.3.7
flask-login
Flask-SQLAlchemy
Flask-Mail